refactor: 클래스 연습 클래스를 연습 패키지로 이동
