+++ src/test/java/study/practice/StringTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.practice;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
